README.md
This repository contains the quizzes along with their SQL commands for the concepts taught in the course SQL for Data Analysis by MODE at Udacity. 
   

NOTE: The data dump file and the above commands works well with PostgreSQL (may or may not work with MySQL and others).

Now, SQL commands can be used to explore the dataset and run queries on it.